+++ src/CheckPlayerStatus.js
