+++ src/Model/BridgeGame.js
@@ -1,6 +1,6 @@
-const { makeBridge } = require('./BridgeMaker');
-const { generate } = require('./BridgeRandomNumberGenerator');
-const OutputMaker = require('./Utils/OutputMaker');
+const { makeBridge } = require('../BridgeMaker');
+const { generate } = require('../BridgeRandomNumberGenerator');
+const OutputMaker = require('../OutputMaker');
 /**
  * 다리 건너기 게임을 관리하는 클래스
  */
